Added new file for configuration.
Moved all the configuration information into a new file.
